+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## [Unreleased]
 ### Changed
-- `<sub` now takes optional `default` parameter
+- `<sub` now takes optional `default` parameter and warns better
 - Now using Soda-ash 0.4.0 (and Semantic UI 2.2.12; update your CSS)
 - Now using re-frame 0.10.1
 - Now using ClojureScript 1.9.908
 ### Added
-- *Sodium core*. Add image, label, search, and button.
+- *Sodium core*. Add button, grid-column, grid-row image, label, list-na, list-item, and search.
 - *Sodium extensions*. Components that specialize the core components. So far:
    app-title, and labelled-field.
 - *re-frame utilities*. Added `vec->fn` and `event->fn`.